Create util for POST api/groups
