+++ fwb/app/api/groups/route.ts
@@ -1,6 +1,9 @@
 import { NextRequest, NextResponse } from "next/server";
 import { auth, currentUser } from "@clerk/nextjs";
 import supabaseClient from "@/supabase";
+import { insertGroup } from  "@/app/api/groups/utils/group_utils";
+
+
 interface Group {
   created_at: Date;
   name: string;
@@ -11,6 +14,10 @@ interface Group {
   // have to work out how to give admin permissions within groups
 }
 
+// Create a new group
+export async function POST(request: NextRequest, response: NextResponse) {
+  return (await insertGroup(request));
+}
 // Get all groups
 export async function GET(request: NextRequest & { query: Record<string, string> }, response: NextResponse) {
   try {
@@ -22,44 +29,6 @@ export async function GET(request: NextRequest & { query: Record<string, string>
     }
 
     return NextResponse.json({ success: true, data }, { status: 200 });
-  } catch (error) {
-    return NextResponse.json({ error: "Internal Server Error" }, { status: 500 });
-  }
-}
-// Create a new group
-export async function POST(request: NextRequest, response: NextResponse) {
-  try {
-    const { userId } = auth();
-    const user = await currentUser();
-
-    if (userId && user) {
-      const formData = await request.formData();
-      const newGroup = {
-        name: formData.get("name"),
-        users: formData.get("users"), // Array in the form
-        discounts: formData.get("discounts"), // Array in the form
-        admins: formData.get("admins"),
-        public: formData.get("public"),
-      };
-
-      // Create a Supabase client with the current user's access token
-      const token = request.headers.get("supabase_jwt");
-      if (!token) {
-        return NextResponse.json({ error: "Could not create supabase access token" }, { status: 401 });
-      }
-      const supabase = await supabaseClient(token);
-
-      // Insert the new discount into the database
-      const { data, error } = await supabase.from("groups").insert([newGroup]).select();
-
-      if (error) {
-        return NextResponse.json({ error: "Failed to create group" }, { status: 500 });
-      }
-
-      return NextResponse.json({ success: true, data }, { status: 200 });
-    } else {
-      return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
-    }
   } catch (error) {
     return NextResponse.json({ error: "Internal Server Error" }, { status: 500 });
   }
